feat: migrate to TypeScript
