+++ .vscode/launch.json
@@ -1,0 +1,26 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Debug Program",
+      "sourceMaps": true,
+      "program": "${workspaceFolder}/dist/index.js",
+      //"program": "${workspaceFolder}/src/index.ts",
+      //"preLaunchTask": "tsc: build - tsconfig.json",
+      //"preLaunchTask": "npm: debug",
+      "outFiles": [
+        "${workspaceFolder}/dist/**.js"
+      ],
+      "args": [
+        "Darcy Wong", 
+        ".\\RAD\\MEP-Q2-2020.json", 
+        "C:\\My Paratext 8 Projects\\MEP"
+      ]
+    }
+  ]
+}
