+++ RAD/MEP-Q2-2020.json
@@ -1,0 +1,23 @@
+{
+    "1CO": {
+        "exegesisAndFirstDraft": 16
+    },
+    "2CO": {
+        "teamChecking": 13
+    },
+    "GAL": {
+        "advisorCheckAndBackTranslation": 6
+    },
+    "EPH": {
+        "communityTesting": 6
+    },
+    "PHP": {
+        "consultantChecking": 4
+    },
+    "COL": {
+        "publish": 4
+    },
+    "1JN": {
+        "exegesisAndFirstDraft": 1
+    }
+}
