+++ package.json
@@ -4,9 +4,10 @@
   "description": "Utility to convert project progress from JSON file to XML",
   "main": "index.js",
   "scripts": {
+    "build": "tsc -p .",
+    "start": "node dist/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/mseag/pt-progress.git"
@@ -23,8 +24,16 @@
   },
   "homepage": "https://github.com/mseag/pt-progress#readme",
   "dependencies": {
+    "commander": "^5.1.0",
+    "path": "^0.12.7",
     "vkbeautify": "^0.99.3",
     "xml-formatter": "^1.2.1",
     "xml2json": "^0.12.0"
+  },
+  "devDependencies": {
+    "@types/node": "^14.0.13",
+    "nodemon": "^2.0.4",
+    "ts-node": "^8.10.2",
+    "typescript": "^3.9.5"
   }
 }
